Conflicts:
	collect-server/pom.xml

+++ collect-server/src/main/java/org/openforis/collect/remoting/service/DataService.java
@@ -5,6 +5,7 @@ package org.openforis.collect.remoting.service;
 
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
@@ -13,6 +14,12 @@ import java.util.Map;
 
 import org.apache.commons.lang3.StringUtils;
 import org.openforis.collect.concurrency.CollectJobManager;
+import org.openforis.collect.event.EventProducer;
+import org.openforis.collect.event.EventQueue;
+import org.openforis.collect.event.RecordDeletedEvent;
+import org.openforis.collect.event.RecordEvent;
+import org.openforis.collect.event.RecordStep;
+import org.openforis.collect.event.RecordTransaction;
 import org.openforis.collect.io.data.BulkRecordMoveJob;
 import org.openforis.collect.manager.CodeListManager;
 import org.openforis.collect.manager.RecordFileManager;
@@ -20,8 +27,8 @@ import org.openforis.collect.manager.RecordIndexException;
 import org.openforis.collect.manager.RecordIndexManager.SearchType;
 import org.openforis.collect.manager.RecordManager;
 import org.openforis.collect.manager.RecordPromoteException;
-import org.openforis.collect.manager.SessionManager;
-import org.openforis.collect.manager.SessionRecordFileManager;
+import org.openforis.collect.manager.RecordSessionManager;
+import org.openforis.collect.manager.SessionEventDispatcher;
 import org.openforis.collect.metamodel.proxy.CodeListItemProxy;
 import org.openforis.collect.model.CollectRecord;
 import org.openforis.collect.model.CollectRecord.Step;
@@ -69,7 +76,7 @@ import org.springframework.transaction.annotation.Transactional;
 public class DataService {
 	
 	@Autowired
-	private SessionManager sessionManager;
+	private RecordSessionManager sessionManager;
 	@Autowired
 	private transient RecordManager recordManager;
 	@Autowired
@@ -77,11 +84,13 @@ public class DataService {
 	@Autowired
 	private transient RecordFileManager fileManager;
 	@Autowired
-	private transient SessionRecordFileManager sessionFileManager;
-	@Autowired
 	private transient RecordIndexService recordIndexService;
 	@Autowired
 	private transient CollectJobManager collectJobManager;
+	@Autowired
+	private transient SessionEventDispatcher sessionEventDispatcher;
+	@Autowired
+	private transient EventQueue eventQueue;
 	
 	/**
 	 * it's true when the root entity definition of the record in session has some nodes with the "collect:index" annotation
@@ -100,7 +109,6 @@ public class DataService {
 		Step step = stepNumber == null ? null: Step.valueOf(stepNumber);
 		CollectRecord record = recordManager.checkout(survey, user, id, step, sessionState.getSessionId(), forceUnlock);
 		sessionManager.setActiveRecord(record);
-		sessionFileManager.resetTempInfo();
 		prepareRecordIndexing();
 		Locale locale = sessionState.getLocale();
 		return new RecordProxy(record, locale);
@@ -152,21 +160,25 @@ public class DataService {
 		return result;
 	}
 
-	@Transactional
 	@Secured("ROLE_ENTRY")
 	public RecordProxy createRecord(String rootEntityName, String versionName, Step recordStep) throws RecordPersistenceException, RecordIndexException {
 		SessionState sessionState = sessionManager.getSessionState();
 		if ( sessionState.isActiveRecordBeingEdited() ) {
 			throw new MultipleEditException();
 		}
-		String sessionId = sessionState.getSessionId();
 		CollectSurvey activeSurvey = sessionState.getActiveSurvey();
 		User user = sessionState.getUser();
-		CollectRecord record = recordManager.create(activeSurvey, rootEntityName, user, versionName, sessionId, recordStep);
+		CollectRecord record = recordManager.instantiateRecord(activeSurvey, rootEntityName, user, versionName, recordStep);
+		NodeChangeSet changeSet = recordManager.initializeRecord(record);
+
+		List<RecordEvent> events = new EventProducer().produceFor(changeSet, user.getName());
+		sessionManager.onEvents(events);
+		
 		sessionManager.setActiveRecord(record);
 		prepareRecordIndexing();
-		Locale locale = sessionState.getLocale();
-		RecordProxy recordProxy = new RecordProxy(record, locale);
+		
+		
+		RecordProxy recordProxy = new RecordProxy(record, sessionState.getLocale());
 		return recordProxy;
 	}
 	
@@ -175,12 +187,18 @@ public class DataService {
 	public void deleteRecord(int id) throws RecordPersistenceException {
 		SessionState sessionState = sessionManager.getSessionState();
 		CollectSurvey survey = sessionState.getActiveSurvey();
-		//TODO check that the record is in ENTRY phase: only delete in ENTRY phase is allowed
-		CollectRecord record = recordManager.load(survey, id, Step.ENTRY);
+		String userName = sessionState.getUser().getName();
+
+		CollectRecord record = recordManager.load(survey, id);
+		if (record.getStep() != Step.ENTRY) {
+			throw new IllegalStateException("Cannot delete a record not in ENTRY phase");
+		}
 		fileManager.deleteAllFiles(record);
 		recordManager.delete(id);
+		
+		publishRecordDeletedEvent(record, record.getStep().toRecordStep(), userName);
 	}
-	
+
 	@Transactional
 	@Secured("ROLE_ENTRY")
 	public void saveActiveRecord() throws RecordPersistenceException, RecordIndexException {
@@ -193,12 +211,13 @@ public class DataService {
 		record.setOwner(user);
 		String sessionId = sessionState.getSessionId();
 		recordManager.save(record, sessionId);
-		if ( sessionFileManager.commitChanges(record) ) {
+		if ( sessionManager.commitRecordFileChanges(record) ) {
 			recordManager.save(record, sessionId);
 		}
 		if ( isCurrentRecordIndexable() ) {
 			recordIndexService.permanentlyIndex(record);
 		}
+		sessionEventDispatcher.recordSaved(record);
 	}
 
 	@Transactional
@@ -206,13 +225,17 @@ public class DataService {
 	public NodeChangeSetProxy updateActiveRecord(NodeUpdateRequestSetProxy requestSet) throws RecordPersistenceException, RecordIndexException {
 		sessionManager.checkIsActiveRecordLocked();
 		CollectRecord activeRecord = getActiveRecord();
-		NodeUpdateRequestSet reqSet = requestSet.toNodeUpdateRequestSet(codeListManager, sessionFileManager, sessionManager, activeRecord);
+		NodeUpdateRequestSet reqSet = requestSet.toNodeUpdateRequestSet(codeListManager, sessionManager, activeRecord);
 		NodeChangeSet changeSet = updateRecord(activeRecord, reqSet);
 		if ( ! changeSet.isEmpty() && isCurrentRecordIndexable() ) {
 			recordIndexService.temporaryIndex(activeRecord);
 		}
-		Locale currentLocale = getCurrentLocale();
-		NodeChangeSetProxy result = new NodeChangeSetProxy(activeRecord, changeSet, currentLocale);
+		
+		String userName = sessionManager.getSessionState().getUser().getName();
+		List<RecordEvent> events = new EventProducer().produceFor(changeSet, userName);
+		sessionManager.onEvents(events);
+		
+		NodeChangeSetProxy result = new NodeChangeSetProxy(activeRecord, changeSet, getCurrentLocale());
 		if ( requestSet.isAutoSave() ) {
 			try {
 				saveActiveRecord();
@@ -224,7 +247,7 @@ public class DataService {
 		return result;
 	}
 
-	protected NodeChangeSet updateRecord(CollectRecord record, NodeUpdateRequestSet nodeUpdateRequestSet) throws RecordPersistenceException, RecordIndexException {
+	private NodeChangeSet updateRecord(CollectRecord record, NodeUpdateRequestSet nodeUpdateRequestSet) throws RecordPersistenceException, RecordIndexException {
 		List<NodeUpdateRequest> opts = nodeUpdateRequestSet.getRequests();
 		NodeChangeMap result = new NodeChangeMap();
 		for (NodeUpdateRequest req : opts) {
@@ -287,26 +310,30 @@ public class DataService {
 		}
 	}
 	
+	@Transactional
 	@Secured("ROLE_ENTRY")
 	public void promoteToCleansing() throws RecordPersistenceException, RecordPromoteException  {
 		promote(Step.CLEANSING);
 	}
 
+	@Transactional
 	@Secured("ROLE_CLEANSING")
 	public void promoteToAnalysis() throws RecordPersistenceException, RecordPromoteException  {
 		promote(Step.ANALYSIS);
 	}
 	
-	@Transactional
 	protected void promote(Step to) throws RecordPersistenceException, RecordPromoteException  {
 		sessionManager.checkIsActiveRecordLocked();
 		SessionState sessionState = sessionManager.getSessionState();
 		CollectRecord record = sessionState.getActiveRecord();
+		String userName = sessionState.getUser().getName();
 		Step currentStep = record.getStep();
 		Step exptectedStep = to.getPrevious();
 		if ( exptectedStep == currentStep ) {
 			User user = sessionState.getUser();
+			sessionEventDispatcher.recordSaved(record);
 			recordManager.promote(record, user);
+			publishRecordPromotedEvents(record, userName);
 			recordManager.releaseLock(record.getId());
 			sessionManager.clearActiveRecord();
 			if ( isCurrentRecordIndexable() ) {
@@ -316,7 +343,7 @@ public class DataService {
 			throw new IllegalStateException("The active record cannot be submitted: it is not in the exptected phase: " + exptectedStep);
 		}
 	}
-	
+
 	protected boolean isRecordIndexEnabled() {
 		return recordIndexService.isInited();
 	}
@@ -325,33 +352,35 @@ public class DataService {
 		return isRecordIndexEnabled() && hasActiveSurveyIndexedNodes;
 	}
 
+	@Transactional
 	@Secured("ROLE_ANALYSIS")
 	public void demoteToCleansing() throws RecordPersistenceException {
 		demote(Step.CLEANSING);
 	}
 	
+	@Transactional
 	@Secured("ROLE_CLEANSING")
 	public void demoteToEntry() throws RecordPersistenceException {
 		demote(Step.ENTRY);
 	}
 		
-	@Transactional
-	protected void demote(Step to) throws RecordPersistenceException {
+	protected void demote(Step toStep) throws RecordPersistenceException {
 		sessionManager.checkIsActiveRecordLocked();
 		SessionState sessionState = sessionManager.getSessionState();
 		CollectRecord record = sessionState.getActiveRecord();
-		Step currentStep = record.getStep();
-		Step exptectedStep = to.getNext();
-		if ( exptectedStep == currentStep ) {
-			CollectSurvey survey = sessionState.getActiveSurvey();
-			User user = sessionState.getUser();
-			Integer recordId = record.getId();
-			recordManager.demote(survey, recordId, record.getStep(), user);
-			recordManager.releaseLock(recordId);
-			sessionManager.clearActiveRecord();
-		} else {
-			throw new IllegalStateException("The active record cannot be demoted: it is not in the exptected phase: " + exptectedStep);
+		Step fromStep = record.getStep();
+		Step exptectedFromStep = toStep.getNext();
+		if ( exptectedFromStep != fromStep ) {
+			throw new IllegalStateException("The active record cannot be demoted: it is not in the exptected phase: " + exptectedFromStep);
 		}
+		CollectSurvey survey = sessionState.getActiveSurvey();
+		String userName = sessionState.getUser().getName();
+		User user = sessionState.getUser();
+		Integer recordId = record.getId();
+		publishRecordDeletedEvent(record, fromStep.toRecordStep(), userName);
+		recordManager.demote(survey, recordId, record.getStep(), user);
+		recordManager.releaseLock(recordId);
+		sessionManager.clearActiveRecord();
 	}
 
 	/**
@@ -460,18 +489,51 @@ public class DataService {
 	}
 	
 	@Secured("ROLE_ADMIN")
-	public SurveyLockingJobProxy moveRecords(String rootEntity, int fromStepNumber, boolean promote) {
+	public SurveyLockingJobProxy moveRecords(String rootEntity, int fromStepNumber, final boolean promote) {
 		BulkRecordMoveJob job = collectJobManager.createJob(BulkRecordMoveJob.class);
 		SessionState sessionState = getSessionState();
+		final String userName = sessionState.getUser().getName();
 		job.setSurvey(sessionState.getActiveSurvey());
 		job.setRootEntity(rootEntity);
 		job.setPromote(promote);
-		job.setFromStep(Step.valueOf(fromStepNumber));
+		final Step fromStep = Step.valueOf(fromStepNumber);
+		job.setFromStep(fromStep);
 		job.setAdminUser(sessionState.getUser());
+		job.setRecordMovedCallback(new BulkRecordMoveJob.Callback() {
+			@Override
+			public void recordMoved(CollectRecord record) {
+				if (promote) {
+					publishRecordPromotedEvents(record, userName);
+				} else {
+					publishRecordDeletedEvent(record, fromStep.toRecordStep(), userName);
+				}
+			}
+		});
 		collectJobManager.startSurveyJob(job);
 		return new SurveyLockingJobProxy(job);
 	}
 
+	private void publishRecordPromotedEvents(CollectRecord record, String userName) {
+		if (! eventQueue.isEnabled()) {
+			return;
+		}
+		List<RecordEvent> events = new EventProducer().produceFor(record, userName);
+		eventQueue.publish(new RecordTransaction(record.getSurvey().getName(), record.getId(), record.getStep().toRecordStep(), events));
+	}
+	
+	private void publishRecordDeletedEvent(CollectRecord record, RecordStep recordStep, String userName) {
+		if (! eventQueue.isEnabled()) {
+			return;
+		}
+		Entity rootEntity = record.getRootEntity();
+		EntityDefinition rootEntityDef = rootEntity.getDefinition();
+		List<RecordDeletedEvent> events = Arrays.asList(new RecordDeletedEvent(record.getSurvey().getName(), 
+				record.getId(), recordStep, String.valueOf(rootEntityDef.getId()), 
+				String.valueOf(rootEntity.getInternalId()), new Date(), userName));
+		String surveyName = record.getSurvey().getName();
+		eventQueue.publish(new RecordTransaction(surveyName, record.getId(), recordStep, events));
+	}
+	
 	protected CollectRecord getActiveRecord() {
 		SessionState sessionState = getSessionState();
 		CollectRecord activeRecord = sessionState.getActiveRecord();
@@ -489,7 +551,7 @@ public class DataService {
 		return sessionState;
 	}
 
-	protected SessionManager getSessionManager() {
+	protected RecordSessionManager getSessionManager() {
 		return sessionManager;
 	}
